+++ src/helpers/string_helpers.ts
@@ -43,7 +43,8 @@ export const MEDIA_TYPES: mastodon.v1.MediaAttachmentType[] = [
     MediaCategory.IMAGE,
 ];
 
-
+// @something
+export const at = (str: string): string => str.startsWith('@') ? str : `@${str}`;
 // [Bracketed]
 export const bracketed = (str: string): string => str.startsWith('[') ? str : `[${str}]`;
 // Prefix a string with [Brackets] and a space